requirements.txt
torch==2.7.0
torchvision==0.22.0
--extra-index-url https://download.pytorch.org/whl/cu126
unsloth==2025.5.7
PyYAML==6.0.2
pandas==2.2.3
transformers==4.51.3
Jinja2==3.0.2
rich==14.0.0